Create a new ChartsUseCase and a ChartsView (used in History & Forecast) for reusability purposes
