+++ app/src/main/java/com/weatherxm/usecases/ChartsUseCase.kt
@@ -1,0 +1,12 @@
+package com.weatherxm.usecases
+
+import com.weatherxm.data.HourlyWeather
+import com.weatherxm.ui.common.Charts
+import java.time.LocalDate
+
+interface ChartsUseCase {
+    fun createHourlyCharts(
+        date: LocalDate,
+        hourlyWeatherData: List<HourlyWeather>
+    ): Charts
+}
